Add WebSocketClient folder enhancements
