+++ webapp/src/clients/websocket/constants.ts
@@ -6,7 +6,13 @@ export const WEBSOCKET_EVENT = {
     CLOSE: 'close',
     ERROR: 'error',
     MESSAGE: 'message',
+    EVENT: 'event',
     JOIN: 'join',
     RECONNECT: 'reconnect',
     LEAVE: 'leave',
 } as const;
+
+export const wsMinReconnectRetryTimeMs = 1000; // 1 second
+export const wsReconnectionTimeout = 30000; // 30 seconds
+export const wsReconnectTimeIncrement = 500; // 0.5 seconds
+export const wsPingIntervalMs = 5000; // 5 seconds
