+++ webapp/src/clients/websocket/types.ts
@@ -1,0 +1,8 @@
+// Copyright (c) 2020-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+export enum WebSocketErrorType {
+    Native,
+    Join,
+    ReconnectTimeout,
+}
